Fix warning.
GitOrigin-RevId: 4beca3cc956fcee81cc96beffb1630c11c5b1a98

+++ td/telegram/MessageEntity.cpp
@@ -440,7 +440,7 @@ static vector<Slice> match_bank_card_numbers(Slice str) {
     while (!is_digit(card_number_end[-1])) {
       card_number_end--;
     }
-    auto card_number_size = card_number_end - card_number_begin;
+    auto card_number_size = static_cast<size_t>(card_number_end - card_number_begin);
     if (card_number_size > 2 * digit_count - 1) {
       continue;
     }
